+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add bidirectional conversions between `ruint::Uint` and crypto library `Uint` types behind `ruint` feature toggle. #758
 - Add bidirectional conversions between `Uint` and `u8`, `u16`, `u32`, `u64`, `u128` types. #764
 - Add EDDSA (Ed25519) signature scheme. #757
+- Add constant `Uint::from_uint(..)` function. `Fp::from_fp(..)` is now constant also. #834
 
 ### Changed (Breaking)